Refactor: Improve modularity and readability of the micromet package.
This commit refactors the monolithic `converter.py` module into several smaller, more focused modules to improve modularity and readability.

The key changes are:
- The `AmerifluxDataProcessor` class has been moved to `reader.py`.
- The `Reformatter` class now resides in `reformatter.py` and acts as a pipeline orchestrator.
- All data transformation logic has been extracted into standalone functions in `transformers.py`.
- Hardcoded station information has been moved to `station_info.py`.
- Generic helper functions have been moved to `utils.py`.

Additionally, this commit includes:
- New tests for the refactored code to ensure correctness.
- Fixes for existing broken tests.
- A fix for a bug in the `tau_fixer` function.
- Resolution of a circular import issue.
