+++ tests_micromet/data/sample_ameriflux.dat
@@ -1,0 +1,4 @@
+"TIMESTAMP_START","TIMESTAMP_END","VAR_1","VAR_2"
+"202401010000","202401010030",1.1,2.2
+"202401010030","202401010100",1.2,2.3
+"202401010100","202401010130",-9999,2.4
