+++ tests_micromet/data/sample_toa5.dat
@@ -1,0 +1,7 @@
+"TOA5","CR6","CR6","1056","CR6.Std.09.02","CPU:MicroMet.CR6","50525","Flux_AmeriFluxFormat"
+"TIMESTAMP","RECORD","STAT_1","STAT_2"
+"TS","RN","",""
+"","","Smp","Smp"
+"2024-06-19 12:00:00",1,10.1,20.2
+"2024-06-19 12:30:00",2,11.1,21.2
+"2024-06-19 13:00:00",3,-9999,22.2
